refactor: create tail differently
Prior, tails were drawn by creating a circle and 'carving out' regions
to get a tail. Unfortunately, this 'carving out' was really only
possible by overlaying other shapes that were the same color as the
background. This requires matching of the background, which in some
instances was impossible (particularly if messages were atop one
another).

This fixes this by creating a custom polygon.